Game of Chance Hand In
I couldn't find the .jar file for whatever reason. When I tried to find the artifact file under "out" it was not there. When I tried deleting the jar file in IntelliJ  and making a new one it said a jar file already existed under MANIFEST.MF but I couldn't open that file.